refactor cli k8sworker
